requirements.txt
beautifulsoup4>=4.12.0
chromadb>=1.5.4
huggingface_hub==0.19.4
langchain-text-splitters>=0.0.1
pdfplumber>=0.10.0
pydantic>=2.0.0
pypdf==3.17.4
python-dotenv==1.0.0
rank-bm25==0.2.2
requests>=2.31.0
sentence-transformers==2.2.2
